@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.6 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -34,18 +34,17 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   # agora_uikit: ^1.0.3
-  agora_rtc_engine: ^5.2.0
-  cloud_firestore: ^3.1.2
-  cupertino_icons: ^1.0.2
-  firebase_core: ^1.14.1
-  firebase_core_platform_interface: 4.5.1
-  http: ^0.13.4
-  lottie: ^1.2.2
+  agora_rtc_engine: ^6.2.2
+  firebase_core: ^2.17.0
+  cloud_firestore: ^4.9.3
+  cupertino_icons: ^1.0.6
+  http: ^1.1.0
+  lottie: ^2.6.0
   permission_handler: 
-  provider: ^6.0.2
-  shared_preferences: ^2.0.12
-  syncfusion_flutter_calendar: ^19.4.52
-  get: ^4.6.5
+  provider: ^6.0.5
+  shared_preferences: ^2.2.1
+  syncfusion_flutter_calendar: ^23.1.38
+  get: ^4.6.6
 
 dev_dependencies:
   flutter_test:
@@ -56,7 +55,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^1.0.0
+  flutter_lints: ^2.0.3
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
